@@ -12,6 +12,7 @@ description: Flutter version of Space Mapper with 2020 upgrades
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 version: 3.1.0+19
+publish_to: none
 
 environment:
   sdk: ">=2.12.0 <3.0.0"
@@ -26,7 +27,7 @@ dependencies:
   background_fetch: ^1.0.1  
   http: ^0.13.4
   flutter_map: ^0.14.0
-  webview_flutter: ^2.1.1
+  webview_flutter: ^2.3.1
   sqflite: ^2.0.0+4
   path_provider: ^2.0.5
   share: ^2.0.4
