@@ -1,4 +1,3 @@
-import 'package:asm/ui/web_view.dart';
 import 'package:flutter/material.dart';
 
 import '../app_localizations.dart';
@@ -6,8 +5,9 @@ import '../components/banner_image.dart';
 import '../components/survey_tile.dart';
 import '../mocks/mock_survey.dart';
 import '../models/survey.dart';
-import '../ui/web_view.dart';
 import '../styles.dart';
+
+import 'survey_detail.dart';
 
 const ListItemHeight = 245.0;
 
@@ -71,7 +71,7 @@ class _AvailableSurveysScreenState extends State<AvailableSurveysScreen> {
   Widget _listViewItemBuilder(BuildContext context, int index) {
     final survey = this.surveys[index];
     return GestureDetector(
-      onTap: () => _navigationToLocationDetail(context, survey.webUrl),
+      onTap: () => _navigationToLocationDetail(context, survey.id),
       child: Container(
         height: ListItemHeight,
         child: Stack(
@@ -84,18 +84,17 @@ class _AvailableSurveysScreenState extends State<AvailableSurveysScreen> {
     );
   }
 
-  void _navigationToLocationDetail(BuildContext context, String surveyWebUrl) {
+  void _navigationToLocationDetail(BuildContext context, int surveyID) {
     Navigator.push(context,
-        MaterialPageRoute(builder: (context) => MyWebView(surveyWebUrl)));
+        MaterialPageRoute(builder: (context) => SurveyDetail(surveyID)));
   }
 
   Widget _tileFooter(Survey survey) {
-    final info = SurveyTile(survey: survey, darkTheme: true);
     final overlay = Container(
       padding: EdgeInsets.symmetric(
           vertical: 5.0, horizontal: Styles.horizontalPaddingDefault),
       decoration: BoxDecoration(color: Colors.black.withOpacity(0.5)),
-      child: info,
+      child: SurveyTile(survey: survey, darkTheme: true),
     );
     return Column(
       mainAxisAlignment: MainAxisAlignment.end,
